hashable

+++ src/Bce/BlockChain.hs
@@ -3,16 +3,18 @@ module Bce.BlockChain where
 import Data.Maybe
 import qualified Data.List as DList
 import Bce.Crypto
-import Bce.Hash    
+import Bce.Hash
+import GHC.Int(Int64)     
 
-type TimeStamp = Int
+type TimeStamp = Int64
 
-data TxOutput = TxOutput { outputAmount :: Double
+data TxOutput = TxOutput { outputAmount :: Int64
                          , outputPubKey :: PubKey
                          } deriving (Show, Eq)
+
 data TxOutputRef = TxOutputRef {
       outputRefTxId :: Hash
-    , outputRefOutputIdx :: Int
+    , outputRefOutputIdx :: Int64
       } deriving (Show, Eq)
 
 
@@ -20,16 +22,16 @@ data TxInput = TxInput { inputOutputRef :: TxOutputRef } deriving (Show, Eq)
 
 data Transaction =
     CoinbaseTransaction {
-      txOutputs :: [TxOutput] }
+    txOutputs :: [TxOutput] }
   | Transaction {
-        txInputs :: [TxInput]
-      , txOutputs :: [TxOutput]
-      , txSignature :: Hash } deriving (Show, Eq)
+      txInputs :: [TxInput]
+    , txOutputs :: [TxOutput]
+    , txSignature :: Hash } deriving (Show, Eq)
 
 data BlockHeader = BlockHeader {
       bhTransactionsHash :: Hash
     , bhPrevBlockHeaderHash :: Hash
-    , bhNonce :: Int
+    , bhNonce :: Int64
     , bhWallClockTime :: TimeStamp
     } deriving (Show, Eq)
 
@@ -37,3 +39,4 @@ data Block = Block { blockHeader:: BlockHeader
                    , blockTransactions :: [Transaction] } deriving (Show, Eq)
 
 data BlockChain = BlockChain { blockChainBlocks :: [Block] } deriving (Show, Eq)
+
